requirements.txt
google-adk[a2a]
python-dotenv
uvicorn
google-adk